chore: change js extension to ts
