+++ package.json
@@ -46,7 +46,9 @@
     "typescript": "^4.2.4",
     "url-parse": "^1.5.1",
     "wait-on": "^5.3.0",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^1.0.1",
+    "ra-core": "^3.15.1",
+    "graphql-tag": "^2.12.0"
   },
   "devDependencies": {
     "@reactioncommerce/eslint-config": "^2.2.0",
