+++ src/components/ProductList.js
@@ -7,21 +7,23 @@ import {
   TextField,
   EditButton
 } from "react-admin";
-import ProductFilter from "./ProductFilter";
 
-export default class ProductList extends React.Component {
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-  render() {
-    return (
-      <List {...this.props} filters={<ProductFilter />}>
-        <Datagrid>
-          <TextField source="title" lable="Product"/>
-          <TextField source="description" />
-          <TextField source="pricing.displayPrice" label="Price"/>
-          <BooleanField source="isVisible" />
-          <EditButton />
-          <ShowButton />
-        </Datagrid>
-      </List>);
-  }
+/**
+ * List of all products.
+ * @param {object} props Products data.
+ * @returns {Component} Component listing all products.
+ */
+export function ProductList(props) {
+  return (
+    <List {...props} >
+      <Datagrid>
+        <TextField source="title" lable="Product"/>
+        <TextField source="description" />
+        <TextField source="pricing.displayPrice" label="Price"/>
+        <BooleanField source="isVisible" />
+        <EditButton />
+        <ShowButton />
+      </Datagrid>
+    </List>);
 }
+
