+++ src/components/ProductView.js
@@ -9,21 +9,24 @@ import {
 // eslint-disable-next-line react/prop-types
 const ProductTitle = ({ record }) => <span>{record.title}</span>;
 
-export class ProductView extends React.Component {
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-  render() {
-    return (
-      <Show title={<ProductTitle/>} {...this.props}>
-        <SimpleShowLayout>
-          <TextField source="title" lable="Product"/>
-          <TextField source="slug" label="Permalink"/>
-          <TextField source="pageTitle" label="Subtitle"/>
-          <TextField source="vendor" />
-          <TextField source="description" />
-          <TextField source="originCountry" label="Origin Country"/>
-          <BooleanField source="shouldAppearInSitemap" label="Included in sitemap?"/>
-        </SimpleShowLayout>
-      </Show>
-    );
-  }
+/**
+ * Displays single product.
+ * @param {object} props Product data.
+ * @returns {Component} Component listing one product.
+ */
+export function ProductView(props) {
+  return (
+    <Show title={<ProductTitle/>} {...props}>
+      <SimpleShowLayout>
+        <TextField source="title" lable="Product"/>
+        <TextField source="slug" label="Permalink"/>
+        <TextField source="pageTitle" label="Subtitle"/>
+        <TextField source="vendor" />
+        <TextField source="description" />
+        <TextField source="originCountry" label="Origin Country"/>
+        <BooleanField source="shouldAppearInSitemap" label="Included in sitemap?"/>
+      </SimpleShowLayout>
+    </Show>
+  );
 }
+
